projects/esgithub/pyproject.toml
[project]
name = "esgithub"
version = "0.1.0"
description = "Search your issues and pull requests with embedding."
readme = "README.md"
requires-python = "==3.13.2"
dependencies = [
  "numpy==2.2.3",
  "openai==1.66.3",
  "platformdirs==4.3.6",
  "pygithub==2.6.1",
  "tomli==2.2.1",
  "tqdm==4.67.1",
  "typer==0.15.2",
]
